Revert "Merge branch 'snake-construction' into main"
This reverts commit 7102e3184b4c25b101843522268f6c631630eb42.

I thought units having a relative position to head is a good idea for 
not updating their positions every frame. However, I noticed that for a 
long and frequent turning snake, relative positions of units have to be 
updated frequently.

Also for a long and frequent turning snake, a relative position of a 
tail unit to head doesn’t seem to have a meaning. This unit is like 
an independent entity now.

I am returning to the first idea of units with independent positions.
This also leads to not calculating real positions of units before 
drawing in the Draw method (although is just a summation).