Fix stack overflow problem
The problem was forgetting that I'm in unit.draw method and call the 
same method in itself. Because of infinite recursion, stack overflow 
occurred.

The problem was not about whether the snake is on the right edge of 
the screen. Just the opposite if the unit was not on any edge or 
corner unit.draw function was called again.