Implement snake rotation
Change snake and unit structure. Snake now holds pointers to head and 
tail units. Units are structured as doubly linked list.

When a rotation input is pressed, a new unit with 0 length and the new 
direction is created and assigned as the head unit. 

While moving the head unit and increasing its length, decrease length 
of the tail unit.

When the length of tail unit is 0 or negative, remove it.