Remove screenRect struct
Remove unnecessary screenRect struct and make screenRect.draw method a 
function.

Creating a screenRect object and then calling its draw method instantly
was unnecessary and not elegant.

I was hoping this change to decrease the number of lines in the 
assembly code. In total it only decreased by 1 line. For (*unit).draw 
method and draw function(old screenRect.draw method), there are no 
changes in the number of lines.