Prevent taking the same turns rapidly.
Making the snake turn left or turn right twice rapidly makes the snake 
rotate into its own body. This had to be prevented so players can make 
these kinds of rapid turns in the late game.

Solve this problem by pushing the second turn into the queue and taking
this turn when the head moves a safe distance.