Make collision check compatible with teleporting.
In some cases where some units of the snake are teleported, the 
collision check was not working. Solve this issue by dividing 
rectangles at screen edges, then check these smaller rectangles if they
overlap. Create a method named slice for this dividing task.

Skip unit that is next to the head while checking collision, since the 
intersection of head and next unit is impossible.

Change name of screenRect struct to rectF64.