Fix false collision on screen edges
While the snake moves on one side of the screen, turning the snake to 
the side where the snake is smaller than the other causes a collision.

Solve this problem by loosening the intersection detection between the 
snake's head and its second unit.

Increase game size on the GitHub pages site.