Fix another false collision problem
Occurs when the snake head is on a screen edge and the player turns the
snake in both directions in a row. Resolve this and previous false 
collisions by increasing the collision tolerance when the second unit 
is on a screen edge.

Increase the game size and make it equal to the screen size.