Make shaders compatible with teleportation
Previously, while a unit was teleporting, its split rectangles are 
drawn as independent rectangles. So when using the round shader, a 
teleporting unit was drawn as two stadium graphical shapes.

Solve this issue by adding "position in unit" info to the rect struct. 
Also, send the total unit length to round and hollow shaders.