Move rectangle creation in draw method to the update
Have separate rectangles in the unit for collision checking and drawing.
This makes it possible to move drawable rectangle creation to the 
update method.

Remove the slicer interface as there are no longer common rectangles 
for drawable and collidable interfaces.

Remove the snake.draw method and draw the snake in the Game.draw method.